Impr: adds refactoring for authentication and behaving accordingly when logging in
