+++ Core/functions.php
@@ -45,22 +45,10 @@ function view($path, $headerArray = [])     //  function for files in views dir
     return basePath('views/' . $path);
 }
 
-function login($user)
+function redirect($path)
 {
-    $_SESSION['user'] = [
-        'email' => $user['email'],
-    ];
-
-    session_regenerate_id(true);
-}
-
-function logout()
-{
-    $_SESSION = [];
-    session_destroy();
-
-    $params = session_get_cookie_params();
-    setcookie('PHPSESSID', '', time() - 3600, $params['path'], $params['domain'], $params['secure'], $params['httponly']);
+    header("Location: $path");
+    exit();
 }
